save versions

+++ src/gh.py
@@ -1,27 +1,31 @@
-import os, sys, shutil, logging, re, urllib.request, pathlib
+import os, sys, shutil, logging, re, urllib.request, pathlib, json
 from github import Github
 
 class GH():
-    def __init__(self, ghToken, update):
+    def __init__(self, ghToken, updateAll):
         self.token = ghToken
-        self.update = update
+        self.updateAll = updateAll
         self.github = Github(self.token)
 
     def downloadReleaseAssets(self, module):
+        
+        fpath = f"./base/{module['repo']}/"
+        update=self.updateAll
+        
+        existed = os.path.isdir(fpath)
         
         if 'local' in module and module['local']:
             shutil.copytree(f"./asset/{module['repo']}", f"./base/{module['repo']}", dirs_exist_ok=True)
             return True
         
-        fpath = f"./base/{module['repo']}/"
-        if os.path.isdir(fpath) and not self.update:
-            logging.info(f"[{module['repo']}] existed")
-            return True
-        
         pathlib.Path(fpath).mkdir(parents=True, exist_ok=True)
         
         if 'link' in module:
-            logging.info(f"[{module['repo']}] Downloading: {module['link']}")
+            if existed and not update:
+                logging.info(f"[{module['key']}] existed")
+                return True
+            
+            logging.info(f"[{module['key']}] Downloading: {module['link']}")
             req = urllib.request.Request(module['link'], headers={'User-Agent': 'Mozilla/5.0'})
             response = urllib.request.urlopen(req)
             with open(fpath+f"{module['filename']}", 'wb') as f:
@@ -39,16 +43,38 @@ class GH():
             logging.warning(f"No available release for: {module['repo']}")
             return
         ghLatestRelease = releases[0]
-
+        latestTag = ghLatestRelease.tag_name
+        
+        existedTag = module['tag']
+        
+        if not update and latestTag != existedTag:
+            print(module['key']+":\n")
+            needUpdate = input(f"update {module['repo']} from <{existedTag}> to <{latestTag}>?")
+            if needUpdate == '':
+                update = True
+        
+        need = latestTag != existedTag 
+                
+        if existed and (not need or not update):
+            logging.info(f"[{module['key']}] existed")
+            return True
+        
         for pattern in module["regex"]:
             found = False
             for asset in ghLatestRelease.get_assets():
                 if re.search(pattern, asset.name):
                     found = True
-                    logging.info(f"[{module['repo']}] Downloading: {asset.name}")
+                    logging.info(f"[{module['key']}] Downloading: {asset.name}")
                     urllib.request.urlretrieve(asset.browser_download_url, f"{fpath}{asset.name}")
             if not found:
                 logging.warning(f"{module['regex']} not found in {module['repo']}")
                 sys.exit(1)
                 return
+        
+        with open('./settings.json', 'r') as f:
+            settings = json.load(f)
+        settings["moduleList"][module["key"]]["tag"] = latestTag
+        with open('./settings.json', 'w') as f:
+            json.dump(settings, f, indent=4)
+            
         return True
